FBD/requirements.txt
pandas>=1.3.0
python-dotenv>=0.19.0
requests>=2.25.1
obonet>=0.4.0


